% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgm.R
\name{cgm}
\alias{cgm}
\title{Centering At The Grand Mean}
\usage{
cgm(var, id)
}
\arguments{
\item{var}{The variable to be centered.}

\item{id}{The ID variable.}
}
\value{
The centered variable.
}
\description{
This function grand mean centers level 2 variables in multilevel data. cgm stands for centered at the grand mean.
}
\examples{
var_L2_cgm <- cgm(var = ml_example$var_L2, id = ml_example$ID)

}
